my-api/requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.31
flask-cors==3.1.1
